+++ docs/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- *Administration* tab in Dashboard shows number of jobs using the particular job type or infrastructure.
+  This helps to indicate which job types are actually in use and which can be deleted.
+  It also shows which job types and infrastructures are provided by which plugins in a tree view.
+  ([#392](https://github.com/TheRacetrack/racetrack/issues/392))
+
 ### Changed
 - Job type plugins can now build job images from a single Dockerfile.
   Base Dockerfile has been merged with a job template.